Modbus: Split common part of modbus and thread part.
