+++ src/App/CMakeLists.txt
@@ -19,7 +19,7 @@ target_link_libraries(${APP_NAME} PRIVATE component)
 target_link_libraries(${APP_NAME} PRIVATE Views)
 target_link_libraries(${APP_NAME} PRIVATE Controllers)
 target_link_libraries(${APP_NAME} PRIVATE Model)
-target_link_libraries(${APP_NAME} PRIVATE Modbus)
+target_link_libraries(${APP_NAME} PRIVATE ModbusThread)
 target_link_libraries(${APP_NAME} PRIVATE IO)
 target_link_libraries(${APP_NAME} PRIVATE Configuration)
