+++ src/CMakeLists.txt
@@ -7,5 +7,5 @@ add_subdirectory(Controllers)
 add_subdirectory(FTXUI)
 add_subdirectory(jsoncpp)
 add_subdirectory(io)
-add_subdirectory(Modbus)
+add_subdirectory(ModbusThread)
 add_subdirectory(Common)
