+++ src/ModbusThread/ModbusDaemon.cpp
@@ -2,9 +2,9 @@
 
 #include "Configuration.h"
 #include "FillModel.h"
-#include "modbus.h"
+#include "../ModbusCommon/modbus.h"
 
-namespace Modbus {
+namespace modbusThread {
 
 ModbusDaemon::ModbusDaemon(const std::shared_ptr<controllers::ViewController> &mViewController, const configuration::Configuration &mConfiguration)
   : mViewController(mViewController), mConfiguration(mConfiguration)
@@ -57,10 +57,9 @@ void ModbusDaemon::runFunction()
       break;
     }
     /* It's better to update view from here than update it on every change in model */
-    // TODO: Improve thread communication with using notify.
     mViewController->updateView();
     using namespace std::chrono_literals;
-    std::this_thread::sleep_for(0.1s);
+    std::this_thread::sleep_for(0.25s);
   }
 }
