+++ src/ModbusThread/ModbusSetter.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Configuration.h"
 
-namespace Modbus {
+namespace modbusThread {
 bool setModbus(const configuration::Configuration &modbusConf, int address, int value);
 
 bool setCoilModbus(const configuration::Configuration &modbusConf, uint16_t address, bool value);
